+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.2.0
+
+### Features
+
+- **Border configuration**<br>
+The window border is now configurable to any of the builtin border styles or a custom border, see `:h nvim_open_win()` for the exact specification.
+
+- **Highlight groups configuration**<br>
+The highlight groups of the window are now configurable. It can be set through the `highlights` table in the config and can use any hlgroups there are.
+
 # 1.1.0
 
 ### Features